requirements.txt
azure-keyvault>=0.3.6